fix wrong datatype

+++ proto/src/packets/request_chunk_radius.rs
@@ -3,6 +3,6 @@ use bedrockrs_proto_derive::ProtoCodec;
 
 #[derive(ProtoCodec, Debug, Clone)]
 pub struct RequestChunkRadiusPacket {
-    chunk_radius: VAR<u8>,
+    chunk_radius: VAR<u32>,
     chunk_radius_max: u8,
 }
